fix: :zap: Fix a bug when a procedure return null or undefined.
Previously, when a procedure returned null or undefined the server was crashing because it tried to access property of this null or undefined value. Also I was not sure it was possible to simply return  json(null) the the server now send a json object with a result key that is deconstructe in the client.